refactor: cross-platform support for macOS and Linux
- Add OS guard (darwin/linux-gnu) to install.sh and check.sh
- Replace hardcoded macOS paths with $HOME-relative equivalents
- Wrap macOS-only config (Homebrew, JAVA_HOME, Coursier, VS Code, MacPorts,
  open/dscacheutil/powermetrics aliases) in OSTYPE guards
- Add Linux equivalents for Homebrew, zplug, p10k, Coursier, JAVA_HOME,
  flushdns, localip and cputemp
- Deduplicate oh-my-zsh.sh, powerlevel10k and NVM loading (were sourced twice)
- Remove obsolete commented-out code and MacPorts install note

+++ zprofile
@@ -1,38 +1,41 @@
-export PATH="/usr/local/bin:/usr/local/sbin:~/bin:$PATH"
-export PATH="/usr/local/bin:$PATH"
+# PATH
+export PATH="/usr/local/bin:/usr/local/sbin:$HOME/bin:$PATH"
 export PATH="/usr/bin:$PATH"
-export PATH="/usr/local/bin:$PATH"
 export PATH="$HOME/bin/scripts:$PATH"
 export PATH="$HOME/bin/scripts/string-utils:$PATH"
 export PATH="$PATH:$HOME/.dotnet/tools"
-export PATH="/Applications/Visual Studio Code.app/Contents/Resources/app/bin:$PATH"
+
+# GPG
 GPG_TTY=$(tty)
 export GPG_TTY
 
-#Leiningen
+# Java
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  export JAVA_HOME="/Library/Java/JavaVirtualMachines/jdk-21.jdk/Contents/Home"
+  export PATH="/Applications/Visual Studio Code.app/Contents/Resources/app/bin:$PATH"
+  export PATH="/opt/local/bin:/opt/local/sbin:$PATH"  # MacPorts
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+  for _jvm in \
+    "/usr/lib/jvm/default-java" \
+    "/usr/lib/jvm/java-21-openjdk-amd64" \
+    "/usr/lib/jvm/java-21-openjdk"; do
+    [[ -d "$_jvm" ]] && { export JAVA_HOME="$_jvm"; break; }
+  done
+  unset _jvm
+fi
+[[ -n "${JAVA_HOME:-}" ]] && export PATH="$JAVA_HOME/bin:$PATH"
+
+# Coursier
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  export PATH="$PATH:$HOME/Library/Application Support/Coursier/bin"
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+  export PATH="$PATH:$HOME/.local/share/coursier/bin"
+fi
+
+# Leiningen
 export LEIN_JVM_OPTS="-Daether.dependencyCollector.impl=bf"
 
-#JAVA
-#export JAVA_HOME=$(/usr/libexec/java_home -v 17.0.16) #java 17
-# export JAVA_HOME=$(/usr/libexec/java_home -v 16) #java 16
-# export JAVA_HOME=$(/usr/libexec/java_home -v 1.8.381.09) #java 8
-export JAVA_HOME="/Library/Java/JavaVirtualMachines/jdk-21.jdk/Contents/Home" #java 21
-# export JAVA_HOME="/Library/Java/JavaVirtualMachines/jdk-11.jdk/Contents/Home" #java 11
-# export JAVA_HOME="/Library/Java/JavaVirtualMachines/jdk-16.0.2.jdk/Contents/Home" #java 16
-# export JAVA_HOME="/Library/Java/JavaVirtualMachines/jdk-1.8.jdk/Contents/Home" #java 8
-export PATH=$JAVA_HOME/bin:$PATH
-
-# >>> coursier install directory >>>
-export PATH="$PATH:$HOME/Library/Application Support/Coursier/bin"
-# <<< coursier install directory <<<
-
-# Nvm and Node
-export NVM_DIR="$HOME/.nvm"
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-export NODE_OPTIONS=--openssl-legacy-provider
-
-#nvim
+# nvim
 alias nv='nvim'
 alias nv.='nvim .'
 alias v='vim'
@@ -41,7 +44,7 @@ alias vConfig='nvim ~/.vimrc'
 alias nvconfig='nvim ~/.config/nvim/ -c "cd ~/.config/nvim/"'
 alias cdNvimConfig='cd ~/.config/nvim/'
 
-#tmux
+# tmux
 alias tx='tmux'
 alias tConfig='nvim ~/.tmux.conf'
 alias tConfigReload='tmux source-file ~/.tmux.conf'
@@ -50,64 +53,69 @@ alias txl='tmux ls'
 alias txk='tmux kill-session -t'
 alias txn='tmux new -s'
 
-#kitty
+# kitty
 alias kittytheme='kitty +kitten themes'
 alias nvkitty='nv ~/.config/kitty -c "cd ~/.config/kitty"'
 alias updateKitty='curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin'
 
 # Paths
 alias cdTmuxConfig='cd ~/.config/tmux/'
-alias cdScripts='cd $HOME/bin/scripts'
+alias cdScripts="cd $HOME/bin/scripts"
 
-# ALIAS
+# General aliases
 alias ls='ls -la'
-alias cp='cp -iv'                           # Preferred 'cp' implementation
-alias mv='mv -iv'                           # Preferred 'mv' implementation
-alias mkdir='mkdir -pv'                     # Preferred 'mkdir' implementation
-alias cd..='cd ../'                         # Go back 1 directory level (for fast typers)
-alias ..='cd ../'                           # Go back 1 directory level
-alias ...='cd ../../'                       # Go back 2 directory levels
-alias .3='cd ../../../'                     # Go back 3 directory levels
-alias .4='cd ../../../../'                  # Go back 4 directory levels
-alias .5='cd ../../../../../'               # Go back 5 directory levels
-alias .6='cd ../../../../../../'            # Go back 6 directory levels
-alias f='open -a Finder ./'                 # f: Opens current directory in MacOS Finder
-alias ~="cd ~"                              # ~: Go Home
-alias home="cd ~"                           # home: Go Home
-alias c='clear'                             # c: Clear terminal display
-alias path='echo -e ${PATH//:/\\n}'         # path: Echo all executable Paths
+alias cp='cp -iv'
+alias mv='mv -iv'
+alias mkdir='mkdir -pv'
+alias cd..='cd ../'
+alias ..='cd ../'
+alias ...='cd ../../'
+alias .3='cd ../../../'
+alias .4='cd ../../../../'
+alias .5='cd ../../../../../'
+alias .6='cd ../../../../../../'
+alias ~="cd ~"
+alias home="cd ~"
+alias c='clear'
+alias path='echo -e ${PATH//:/\\n}'
 alias reloadConfig='source ~/.zprofile'
 alias myip='curl https://ipinfo.io/ip'
 alias ports='sudo lsof -i -P | grep LISTEN'
 alias lastJsonModified='ls -t *.json | head -n1'
 alias exitCode='echo "Exit code => " $?'
-
-# macOS specific aliases
-alias resetBluetooh='sudo pkill bluetoothd'
-
-# System shortcuts
-alias flushdns='sudo dscacheutil -flushcache;sudo killall -HUP mDNSResponder;say cache flushed'
-alias sha1='/usr/bin/openssl sha1'
-alias md5='md5 -r'
-alias md5sum='md5 -r'
 alias running_services="sudo lsof -nPi -sTCP:LISTEN"
 alias whatismyipaddress='dig +short myip.opendns.com @resolver1.opendns.com'
-alias localip='ipconfig getifaddr en0'
-alias chrome='open -a "Google Chrome"'
-alias change_to_bash='chsh -s /bin/bash'
+alias sha1='/usr/bin/openssl sha1'
 alias setExecutable='chmod +x '
 alias countFiles='find . -type f | wc -l'
 
-## get top process eating memory
+# Process
 alias psmem='ps aux | sort -nr -k 4'
 alias psmem10='ps aux | sort -nr -k 4 | head -10'
+alias pscpu='ps aux | sort -nr -k 3'
+alias pscpu10='ps aux | sort -nr -k 3 | head -10'
 
-## get top process eating cpu ##
-alias pscpu=' ps aux | sort -nr -k 3 '
-alias pscpu10=' ps aux | sort -nr -k 3 | head -10 '
-alias cputemp='sudo powermetrics --samplers smc |grep -i "CPU die temperature"'
+# macOS-specific aliases
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  alias f='open -a Finder ./'
+  alias chrome='open -a "Google Chrome"'
+  alias change_to_bash='chsh -s /bin/bash'
+  alias resetBluetooth='sudo pkill bluetoothd'
+  alias flushdns='sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder; say cache flushed'
+  alias md5='md5 -r'
+  alias md5sum='md5 -r'
+  alias localip='ipconfig getifaddr en0'
+  alias cputemp='sudo powermetrics --samplers smc | grep -i "CPU die temperature"'
+fi
 
-# GIT alias commands
+# Linux-specific aliases
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+  alias flushdns='sudo resolvectl flush-caches 2>/dev/null || sudo systemd-resolve --flush-caches'
+  alias localip='hostname -I | awk "{print \$1}"'
+  alias cputemp='cat /sys/class/thermal/thermal_zone*/temp 2>/dev/null | awk "{printf \"%.1f°C\n\", \$1/1000}"'
+fi
+
+# GIT
 alias add='git add'
 alias st='git status'
 alias log='git log'
@@ -125,14 +133,7 @@ alias gdelete='git branch -D'
 alias gbnew='git branch -m'
 alias lg='lazygit'
 
-##
-##
-
-# MacPorts Installer addition on 2021-07-30_at_09:03:23: adding an appropriate PATH variable for use with MacPorts.
-export PATH="/opt/local/bin:/opt/local/sbin:$PATH"
-# Finished adapting your PATH environment variable for use with MacPorts.
-
-#Dotnet commands
+# Dotnet
 alias dn='dotnet'
 alias dnb='dotnet clean && dotnet restore && dotnet build --force'
 alias dnr='dotnet run'
@@ -140,43 +141,33 @@ alias dnc='dotnet clean'
 alias dnrt='dotnet restore'
 alias dnt='dotnet test'
 
-#Clojure commands
+# Clojure
 alias cljReplRun='clj -M:repl/conjure'
 
-#Kafka commands
-#alias startKafka='docker-compose -f ~/Docker/composes/kafka.yaml --project-name kafka up -d'
-#alias stopKafka='docker-compose -f ~/Docker/composes/kafka.yaml --project-name kafka down'
+# Kafka
 alias startKafka='confluent local kafka start'
 alias stopKafka='confluent local kafka stop'
 
-#Util functions
-
-#Decode jwt token
-jwt-decode() {
-  jq -R 'split(".") |.[0:2] | map(@base64d) | map(fromjson)' <<< $1
-}
-
-#How to cat a file with colors
-# cat <fileName> | jq
-catFileWithColors() {
-    cat $1 | jq
-}
-
-#CERTIFICATES
-print-certificate-pfx() {
-    openssl pkcs12 -in $1 -info -noout
-}
-
-#NPM
+# NPM
 alias n='npm'
 alias nr='n run'
 alias ni='n install'
 
-#databricks
+# Databricks
 alias dbc='databricks'
 
-#json
+# JSON
 alias flattenJson='jq -c .'
 
-# work-specific config (not tracked by git)
-[[ -f ~/.config/zsh-environment/private/work.zsh ]] && source ~/.config/zsh-environment/private/work.zsh
+# Util functions
+jwt-decode() {
+  jq -R 'split(".") |.[0:2] | map(@base64d) | map(fromjson)' <<< $1
+}
+
+catFileWithColors() {
+  cat $1 | jq
+}
+
+print-certificate-pfx() {
+  openssl pkcs12 -in $1 -info -noout
+}
